removed duplicated route position and consolidated data/positions + added actions for logout
