+++ src/routes/oauth/github/callback/+server.ts
@@ -142,7 +142,7 @@ export const GET: RequestHandler = async (event) => {
 		return new Response(null, {
 			status: 302,
 			headers: {
-				Location: '/portfolios'
+				Location: '/data/portfolios'
 			}
 		});
 	} catch (error) {
