+++ src/routes/oauth/google/callback/+server.ts
@@ -133,7 +133,7 @@ export const GET: RequestHandler = async (event) => {
 		return new Response(null, {
 			status: 302,
 			headers: {
-				Location: "/portfolios"
+				Location: "/data/portfolios"
 			}
 		});
